Fix filled contours when used from external modules (#92)
There were no checks if the CPT was a memory object.  If it is then we cannot free it after checking if it is continuous or not.  Also, we cannot just check for a filename ending in ".cpt" since it will have the memory name. Added more comments as well.