The parsing of +n<length>[<unit>] should not override the default unit for data input
The bug was that +n1c would change the deafult dimension to cm for data input even though --PROJ_LENGTH_UNIT=inch was given.  Now, the length is parsed and converted to inchs but no unit is recorded.  Furthermore, arrowline.sh had some bugs (used - instead of -- to set a default value) and vectors.sh were affected by the +n bug so I just changed its unit so the PS is the same.  The arrowline.ps file needed to be updated as well.