Finalize how BB clipping works for test placed via +c<just>
When computing the coordinates for text via justification codes we want those codes to relate to the rectangular bounding box and not the actual map outline, and we want the "throw away of outside" check deactivated since round-off sometimes catches such on-the-border coordinates.  The text itself is still subject to clipping by the BB unless -N is used as well.