Explain why modern mode needs to know page size for PS
In modern mode, plots are automatically cropped but this is not true if plain PostScript (ps) is requested, in which case a media size is recommended.  This PR adds explanation of this to begin and changes the verbosity level of the error to warning (thus requiring -V to be seen).  Addresses #94.