Change from svn to git in building tarballs (#281)
* Change from svn to git in building tarballs

The CMake system relied on svn to extract the archive used to build tar balls.  This has now been modified to use git archive instead by using a .gitattributes file to exclude the test dir and git-specific files.  This is the same approached used in the terminal 5.4 branch.

* remove some comments

Forgot to remove the temp-solution code.