Slanted annots via -B (#286)
* WIP Slanted annots via -B

* Change syntax to -B+a<angle>

* Move angle to axis list

Also added check for N vs S but still not completed.

* Delete lixo.pdf

Oops

* Add test

Add standard test for 45 and 90 degree slants.  Failing for now.

* Update gmt_plot.c

Getting closer

* Update gmt_plot.c

Finalize slanted I think

* Document slanted annotations

Updated the documentation and stricter error checking on the parsing, plus added the finalized test and its PS.

* Remove unneeded enums from GMT_enum_tick

* Final commit

Added revised -B description to the Cookbook as well, fix check for which axes the +a modifier is valid for, and improve message (which is now a warning).

* Added +ap for y-axis

Updated docs, added new figure showing slanted annotations in the cookbook.