Warn users that sorted data may be required in triangulate
Per issue #279 we have added a check to see if the data are sorted in x, and if not then we issue a warning and recommend users first sort their data.  The problem seems to be either a bug in the Watson algorithm or that the same algorithm requires the data to be sorted but never specified that requirement.  The Shewchuck algorithm does not have this issue.