Add OSX builds and refactor TravisCI build config (#317)
Refactor the TravisCI build configuration to make it easier to add OSX support
and to deploy the docs to Github pages later on. Set different builds in the 
matrix for cron jobs and separate testing and building the docs. This is 
necessary since the tests are known to always fail and we want to deploy the 
docs anyway. The build was always passing even if the docs or tests ended in 
failure because the `cd build; ...; cd..;` command always passes. The return 
value is only taken from the last `cd`.