Start a contributing guide with basic information
The guide is written in Markdown in the `CONTRIBUTING.md` file in the
base directory. This is standard across Github and it will automatically
detect this file and include links to it in issues and pull requests.
The guide has basic info about creating an issue, making a PR, and
getting in contact. It's based on the one I use for my python projects:
https://github.com/fatiando/contributing/blob/master/CONTRIBUTING.md

There is no GMT specific information, which has to be added later, but
it's a start.