Use GMT_SESSION_NAME instead of GMT_PPID
Expert users may wish to supply a unique session name which will allow then to use sub-shells while running modern mode.  Thus, we allow this name to be a string and hence are renaming the GMT_PPID environmental variable to a more generic name.  Under the hood, if no such variable is set we create one as before from the parent process id.