Allow -r to take optional g or p (#356)
In anticipation of longer-format options it works better of the -r option can take explicit directives g (for gridline) and p (for pixel) registration.  The -r with no argument is equivalent to -rp.  This makes it easier as well for developers who wish to issue a -r but using variables g or p as needed.