earthtide specify min interval with -T
Consider output of commands
        gmt earthtide -L52/21 -T2010-01-01T/2010-01-01T00:00:10/0.1s
        gmt earthtide -L52/21 -T2010-01-01T/2010-01-01T00:00:10/1s
        gmt earthtide -L52/21 -T2010-01-01T/2010-01-01T00:00:10/20+n
        gmt earthtide -L52/21 -T2010-01-01T/2010-01-01T00:00:10/25+n

which are very confusing. As the first column is changing, while desired
parameters are sometimes computed correctly, and sometimes not.

This is due to problem that if ttdel2 is less
then half second then fmjd is never changed in subroutines for
computation of Moon/Sun position of tidal deformation.