Parsing of plus symbol to indicate ellipsoidal calculation (#416)
mapproject relies on the sign or increments of (in absence of increment) a leading sign in front of desired unit to switch into flat earth (-) or ellipsoidal (+) distance calculations [Default is great clrcle].  With the advent of modifiers this mechanism could failed, such as +u+e which was seen as an unrecognized modifer +e rather than the arguments to +u.  This is now fixed.