Make the vector pen and fill modifiers work properly (#415)
* Fix interactions of vector parameters for color, pen, and -G -W

The -G -W command line settings determine defaults, but these can be overridden via modifiers +g +p when symbol codes are read from file.  On the command line the modifiers override the defaults.

* Fix symbol fill/pen modifiers

The +g and +p modifier for vectors did not always work as expected, and the documentation was not correct.  Head outline pen is by default the same pen as the stem, except width is 50%.  This pen can be overridden via +ppen.

* remove some unused options

Dont need PS_COMMENTS in the final script