Consider line thickness when clipping lines againts rectangular boundary
This PR is in support of the problem discussed in #402 and follows the solution I outlined earlier today.  Currently, the algorithm is limited to maps with rectangular boundaries as well as lines that do not wrap across 360 degrees.  Once the current PR is tested and found acceptable then I will extend that to the other two cases. This PR should not be accepted without trying and runnign the examples.  It currently solves @kristofkoch example and passes the trimline.sh test.  Please switch to this branch and run tests and comment first.  It currently breaks a dozen or so other scripts, but in a good way: Part of my solution requires map clipping to be on (in psxy, we only turned on clipping for symbols, not lines, since lines where clipped numerically, but this is part of the problems).  With clipping, lines close the the boundary, especially those subparallel or parallel to the border, will be clipped and look thinner than before.  This is correct behavior but "breaks" the tests.