Let surface -Q run without requiring -G that is not used anyway (#455)
The -Q option reports better -R -I combination that optimizes speed and convergence; no grid is created.  It is therefore silly to require the -G option.  This has now been corrected.  At the same time I noticed a bad parsing test for -Am: If given and data is not geographic we complain, however the error message was a cut/paste of another message and thus very confusing.