Default -F did not set key internal bool variable (#489)
* Default -F did not set key internal bool variable

When no -F is given we default to  -Fxyzpqrs, and when z is included we need to set P.want_z_output = true; for proper processing. This was done when -F was processed but not if -F was not given.  Fixes issue #486.

* Cannot implicitly set -Fz if only x,y columns

Now check if -F was set implicity then if no z data or trailing text then we quietly turn off that output column.  If explicitly set then of course it yields an error.  Also found a bad memory allocation issue when only a single column was given as input.

* Remove z from list of outputs when input data has none

Basically, needed to reduce number of outputs by one and shuffle the columns one step to the left to remove z.

* Need to update col types

Cannot really set output col types until we know how many z columns there are and what their types may be (which could be set via -fi).