Enhance mapproject -W to return reference point coordinates (#503)
The -W option is used to obtain a map's width or height.  I have expanded it to take typical reference points (as used when placing embellishments, such as colorbar or roses) so that mapproject can return the geographic coordinates of the reference point.  For instance, if you need to map coordinates of the center of an oblique map, you use -WjCM.