When module + is given, any add-ons to common options should not be issues
When an extended synopsis is requested (e.g., gmt psxy +) we suppress output for the common GMT options (e.g., -R, -J).  However, in many cases we have added an extra sentence of two to follow these, and in many of those cases we forgot to add suitable test to prevent these lines from being printed without the common context.