Don't check if W node equals E node for grids that dont repeat (#524)
SOme global grids claim gridline registration but do not have the repeating values at the periodic boundary because they really are pixel-registered grids masquerading as gridline registrered grids.  In that case we should not check if the node on the west equals the node on the east.
Also, I increased the precision in a warning message to show that a value was changed.