Finalize QR handling in psxy and psxyz (#525)
Finalized the implementation of letting -G set the QR color and allow -W to draw the outline of the quiet zone.  The only remaining issue is whether or not we consider the white square background part of the symbol or not.  I am leaning towards the solution that it is up to the user to ensure a contrasting background, otherwise we have no way of eliminating the white rectangle if we wanted to.  Deals with issue #463.