Initial parsing was vulnerable to mistakes (#530)
The recently revised parsing of -Er+ in blockmedian and blockmode lead to a SEGV that is now fixed.   Also, the new parsing of -L in psmask.c failed when the old syntax -L+pname was given (it thought +p was a modifier).  FInally, psmask source code did not show -F in synopsis or usage.