Avoid passing stride = NULL to io_nc_varm_grdfloat (#541)
netCDF introduced a bug in 4.6.2 and fixed it in 4.6.3.  However, macports just now released 4.6.2 and GMT scripts involving netcdf grids would in some cases crash do to the bug.  A fix is to not pass a stride of NULL (which was meant that netCDF would create unit strides - but that is the bug) and instead pass thje unit strides directly.  Closes # 512.