grdimage and nonpolar azimuthal grids issues
See #521 for context.  The original grid is poorly presented since it is actually a global (in longitude, not latitude) pixel-registered grid saved as a nonglobal gridline-registered grid.  However, converting it to a pixel grid with proper header still failed (int only worked for a full range of latitudes).  This PR addresses the problem with the range of latitude since these nonpolar azimuthal grids with 360 range of longitude must be projected as the full square.  With proper global and pixel registration it now works.  I will see if I can figure out if we can handle to badly designed grid but it is tempting to blame the format.