Let grdgradient remember previous statistics for normalization (#547)
Giant grids and/or maps may require you to tile the final image.  Then, it is important that each grdgradient operation uses the same mean and sigma in the normalization.  The new option -Q simplifies how this works by saving the statistics to a hidden file (-Qc) then subsequent grdgradient calls with -Qr will read that file and optionally use these values (depends on how the modifiers +o and +s to -N are used) in the next calculations.  This prevents a change in iamge appearance across seams.