Address various gcc compiler warnings
This PR fixes a bunch of issues that yielded compiler warnings with gcc but now with clang on OSX.  The types of issues addressed are:
Instances where a bool variable was used for counting (bool error++)
Instances with sprintf printing text that could potentially be longer than the buffer
Statements fall through in switches.  While these are all intentional, I have made sure each instance has a comment to that effect.
Some uninitialized variables