Clarify that grdview -N+gfill does not apply to oblique projections
Since we need to sample the grid along the boundaries, the facade fill option does not yet handle cases where the grid is oblique with respect to the map boundaries (i.e., whne +r is used in -R).  The usage and documentation have been updated to state this.  Related to #553 which also pointed out a problem in reading grids via GDAL.