Let gmt_gdalread parse the grid region without changing -R settings (#556)
The gmt_gdalread function used gmt_parse_common_options function to parse a w/e/s/n text string to obtain the domain values,but this had side-effects as that function was only meant to handle the command-line -R setting (there are other items being set).  Thus, when a GDAL grid was read and an oblique projection and domain had been set with -R...+r, the oblique settings were overridden.  For other grids this did not happen since they do not need to parse a w/e/s/n string.
The fix was to use GMT_Get_Values instead and avoid updating the common -R setting.  For me, all but 27 of 825 test pass, unchanged from before this commit.