Standardize -g to use modfiers and not -|+ as flags (#558)
* Standardize -g to use modfiers and not -|+ as flags

The -g options for detecting data gaps had rarely used feature of giving a signed gap.  Then, the sign of the gap going from previous to current record mattered [normally we use the absolute value of any difference].  However, this use of + can interfere with modifier parsing.  Futhermore, the feature was not documented consistently and not used in any examples.  Finally, the option to require all gap specifications be satisfied (-ga) was incorrectly referred to as -g: in the main documentation.

* Remove unneeded space

No need to escape the space before end bracket.