Improve parsing of modifier text with plus-symbols
Since quotes are not passed into GMT, it can be difficult to know if any combination of +letter is a modifier or not.  To eliminate as many cases as we can, we do this:
For argument to an option that takes modifiers, replace any + symbol that is not followed by one of the allowed modifier letters with ASCII 1.
New: THe second time we find a + followed by a valid modifier we know it cannot be a modifer since a modifier can only be given once.  So +tTitle+test will now handle the second +t sequence as plain text.
Bug: If a string ended in a plus then it got removed.

With these improvements we are less vulnerable to this problem and a few tests could be simplified and a few actually had text ending in double-pus-symbols to compensate for the loss of the last.  Ultimately, the only foolproof way to not have a +letter be interpreted as a modifier is to use the octal code for the plus: \053.