Start connecting to closest neighbor (#579)
While not entirely sure how to fix this algorithm, starting off in the direction with the closest neighbor segment seems to avoid the SEGV crashes that @abanbj found.  The three cases now work for me.  I also lowered the severity of the "Trouble" message to VERBOSE rather than ERROR, since it is not really an error...  Closes #427.