Let x2sys_list -N option have modifier to exclude some pairs
The -N option was sort of explained in the module usage but was wrong in the documentation.  In the process I added a new modifier **+p** to apply the test to the number of crossings in any pair of tracks versus the original test which looked at the total number of crossovers between a track and any other tracks.  Partially addresses #586.