Let surface -Q report findings as INFORMATION, not ERROR
The -Q option is writing requested information to stderr, not error.  Yet, to ensure it was written we treated it as an error, confusing users.  THe solution is to temporarily escalate the verboseity setting so that information messages are written, then reset before we quit.  I also added a note in the case the extended regions exceed +/-90 for latitudes.

This closes #454.