Add MD5 and size checks on remote files (#582)
* WIP experimenting with MD5 checks on remote files

We need e mechanism to refresh a users cache files when the file on the server changes.  Also, we need to be aware if a download failed before completing.  This PR is exploring solutions to these problems.

* Update gmt_remote.c

Dealing with different modification times on OSX and Linux.

* Update comments

Looks like Apple is the only one without st_mtime as it exists under Windows and Linux.  Must instead use st_mtimespec.tc_sec.

* Free used memory and add error if MD5 file is corrupted

Tested the time, hash, and size checks.  Need solution for corrupted M5D table itself.

* Add more debug messages

When run with -Vd the whole sequence should print messages.

* More debug statements

Also print when file is found and age < 1 day.

* Only delete files not in the new list.

Improve checks and add more debug messages.

* Limit md5_refresh to one call

Because of the way GMT registeres inputs etc the dm5_refresh function could be called twice.  Added a variable to prevent wasting time on a repeat call.

* FInalize comments

Just improve the comments and clean up some of the debug print statements.

* fix md5 typo (#589)

* Check if server dir exists

If user has never used remote files we must check if server directory exists and if not create it first.

* Check if no connection

Report all errors when getting the gmt_md5_server.txt file fails.

* Disable auto-download on failure

If we cannot connect, then disable auto-download for the rest of the session.