Allow pscontour and grdcontour to take comma-separated lists
This PR lets both contouring tools accept a comma-separated list of specific contours.  The special case of a single contour needs to be given a trailing comma to not be confused with a contour interval.  This buries the old scheme of giving a single contour by using a leading +, which alwasy looked odd for negative contorus (-C+-10).  It is now -C-10, with its trailng comma.  The old scheme is still supported of course but no longer documented.  Closes #461.