Describe fat line rendinging issue in PostScript
Because the default line rendering algorithim in GHostScript is optimized for speed, very fat lines may appear horrible.  This follows issue $431.  The resolution to this has been to add some more documentation into the Cookbook showing this effect using the simple example (Case 2) that @kristofkoch identified.  THe documentation shows these and points out the remedy.  This closes #431.