Disallow map projections when plotting Mercator IMG subsets (#680)
See #674 for context.  If users extract a subset from an IMG grid and requests it to remain in the spherical Mercator format, then one cannot plot this grid with anything but a Cartesian projection.  This PR adds a new check to prevent such a mix of projections and projected data in the 4 modules that plot grids (grdimage, grdcontour, grdview, grdvector).