Fix how psxy and psxyz handles reading symbol size when -i is also usd
Joaquim pointed out that example_22.sh did not seem to care if -Sci or -Sc was given (the i tells psxy that input sizes are in inches).  This was because of a bug that did not detect that when -Sc is given the default unit should be used, and if that is unchanged by gmt.conf it is CM, hence a scaling should occur.  Now fixed.