Look for required executables and warn if not found (#742)
* Look for required executables and warn if not found

Sine GMT uses the executables gs, gm, ffmpeg, and xdg-open we now look to see if they are present in the path and warn if they are not.  Compilation will continue regardless of result.

* Also check for ogr2ogr and gdal_translate

The first is used deep in GMT while the second is used by psconvert to make geotiffs.

* Only check gs under Unix/Linux

Special hunting in registry under Windows.