Correct parsing of -A for modules using GSHHG data (#768)
* Correct parsing of -A in pscoast and elsewhere

For modules that uses the GSHHG dataset, the -A option offers a special modifier +a for affecting Antarctica: One can select either the ice fron or gorunding line as the coastline, and there are expert settings to either skip all of Antarctica or skip everything but Antarctica.  There were several problems:
It was not possible to select two +a modifier codes simultaneously
The bitflags used internally to know what to do had a few logical errors.
I have updated the internal bits to be compatible with multiple selections and I have fixed various logical tests to use bit-comparisons instead of plain equality.  This closes issue #763 .

* Rename macro to SYN_OPT-Area

since no a common option