Let -Tmin/max/inc behave like array creation in MATLAB, etc. (#806)
The new -T parser is pedantic and will complain if (max-min) is not an integer times inc.  However, in MATLAB and other langulages, creating an array like x = 0:2:11 will result in 0, 2, 3, 4, 10 with no fussing whatsoever.  I have changed the parser to only warn about this (requires -V).  But perhaps it should not even be a warning (i.e., requires -Vl)?