Fix confusion between grdvector and gpsgridder examples
Now that grdvector, if given a map projection, draws geovectors, the option to shrink the vector heads for short vectors (i.e., +nlength) must be given in km.  Some scripts gave the argument in inches, which makes no sense.  The "failures" of the gpsgridder examples were a combination of the lack of shrinking heads (wrong cutoff given), the slightly curved geovectors versus the old straigth lines, and a mismatch of cutoff scales for small and large regions.  This PR fixes those issues.  I also simplified gpsgridder -Cnval (which selects how many eigenvectors to use) to optionally take a val < 1, in which case we assume it is the fraction of eigenvalues to use.  This elminites the need to know or compute the total nmber of eigenvalues; the scripts have been updated and shortened accordingly.