Allow different CPT defaults if using earth|srtm_relief files
The GMT CPT default is rainbow but it is not suitable for DEMs since there is no clean separation at sealevel.  This PR introduces the notion of several GMT CPT defaults.  For now there are the default rainbow, the geo for earth_relief, and the srtm for srtm_relief (no ocean).  All tests that passed before still pass.