Fix a few problems with -Ttmin/tmax/tinc parsing and array creation (#817)
THe new -T parser and creater had a few problems.  THis follows up on the earlier PR #773 . These two problems were identified and corrected:
When calendar time args are detected, we set the corresponding input column data type to ABSTIME.  However, we should not set the corresponding output column to the same unless -fo was not set.
When min and/or max t-value pointers are passed to gmt_create_array we must use those values as tin/tmax when creating the array.
The bugs affected filter1d mostly.