Add -C support to modules requiring an optional cpt to paint symbols (#821)
This PR extends -C (no arguments) under modern mode to mean "use current CPT".  If none exist then the module may return an error (e.g., psxy) or move on to create one (e.g., grdimage).  Note that -C is not a common GMT option so  for instance in psmeca there is a -Z option used for this purpose.