Complete missing script and update PS
The test paintallzs.sh in grd2cpt was not completed so no PS was there and it was not listed in CMakeLists.txt.  I finished the script and added a PS .  Also the apimatgrd.sh script created a PS file by another name, so I renamed it to follow the pattern that scripts and plots have the same basename.  Finally, I deleted testapi_mixmatrix.ps as not referenced by anuthing.