Simplify test scripts gmtest.in and their CMakeLists.txt files
Eliminate special provisions for clsasic versys modernize checks in the three gmtest.in scripts.  Instead, use a simple grep to detect if a script is modern mode (looking for gmt begin) and use that to handle the setting of the session ID.  Flag all known failures with the keyword GMT_KNOWN_FAILURE.  ANy script with that word in it will not be compared with an original (which often is missing) and istead will FAIL (known failure) and not add to the failure count.
THe only desired modification would be a count of known failures at the end, so we are at least reminded when running the tests.