IMprove test reporting, allow per-script rms setting for gm (#865)
See #852.  This PR adds the ability to enter a script-specific GRAPHICSMAGICK_RMS setting to override the one configured by CMake.  The reason for this is that some scripts, when run under Linux, just barely fail against the PostScript originals created under macOS, and the differences are due to roundoff and other things related to compiler differences and what not.  We will identify these false positive failures and elevate the rms setting so they pass the test.