Consider line thickness when clipping lines against rectangular boundaries (#426)
* Consider line thickness when clipping lines againts rectangular boundary

This PR is in support of the problem discussed in #402 and follows the solution I outlined earlier today.  Currently, the algorithm is limited to maps with rectangular boundaries as well as lines that do not wrap across 360 degrees.  Once the current PR is tested and found acceptable then I will extend that to the other two cases. This PR should not be accepted without trying and runnign the examples.  It currently solves @kristofkoch example and passes the trimline.sh test.  Please switch to this branch and run tests and comment first.  It currently breaks a dozen or so other scripts, but in a good way: Part of my solution requires map clipping to be on (in psxy, we only turned on clipping for symbols, not lines, since lines where clipped numerically, but this is part of the problems).  With clipping, lines close the the boundary, especially those subparallel or parallel to the border, will be clipped and look thinner than before.  This is correct behavior but "breaks" the tests.

* Fix the case exemplified by @kristofkoch case 1

Case 1 now works.  However, a side effect is some gridlines that fall along map boundaries and reach a pole point adds a crazy horizontal line.  Will have to debug that one later.  Have not yet looked at the other cases.

* Add -W clip path to pscoast as well

With this update, pscoast passes @kristofkoch's test.

* Laying prep work for dealing with slanting borders

The current clipping of fat lines assumes boundaries are rectangular.  To handle arbitrary boundaries we need to compute the angle of the boundary at the crossing.  This commit adds that capability but is not yet utilizing it.

* Add clipping against nonrectangular borders

One more problem remains after this: Circular borders where there is no w/e/s/n side.

* Update some PS files

THese are minor changes related to now having a clip path in psxy for lines.

* Update gmt_map.c

Adding comments.

* Fix bad pen flags during jumps.

Some full-width jumps in x had bad pen flags; this commit fixes that.

* Avoid clipping of frame drawn manually

Since clipoing is on, drawing a frame manually means half is clipped.  These are ancient drawings so just fudge -R to fix this.

* Update a few PS files

Need new PS files after code changes.

* Re-fix fatline

Forgot the point was to show the problem, not fix it.

* Set central meridian inside img2grd

* Undo central meridian.

* Update gmt_support.c

Fix format statement in error message to prevent crash.

* Update gmt_map.c

When joining two lines we now need to worry about the PSL_CLIP flag.  I forgot this an general_.sh started to fail.