Give more control regarding frame type graph
This PR allows you to change the defaut parameters [7.5%] for the length of the extension needed to add a vector at the end of the axis under MAP_FRAME_TYPE=graph mode.  You can append ,length[unit], where unit may be % (extension to tip of vector is this fraction of the axis length) or give an absolute length (with units c, i, p).  Then we extend the axis by that amount before adding the vector head.