Allow modern mode module names with no args to give usage
Running gmt wiggle to get usage would end with no such module since we are not yet in modern mode.  Now, if a session has not yet started (by gmt begin) and users type the name of the modern module name, they will get the usage message.  If arguments are given then we get the error that you are not in modern mode.