Use mergesort in gmtspatial NN so results are stable (#956)
* Use mergesort in gmtspatial NN so results are stable

We found that the nearest neighbor analysis gave slightly different results on macOS as on Linux, and traced it to the fact that qsort is an unstable algorithm (if given identical items then the order is undefined).  Thus, we switch to mergesort (which is stable) in this application.  Exception is (for now) Windows which appears not to have a built-in mergesort function.

* Add a mergesort.c function

Apparently, some Linux dont have it

* Update mergesort.c

Remove reference to util.h

* Update mergesort.c

* Keep trying

See if we can get it to work on Travis.