Disallow -it and check if trailing text is not NULL (#1047)
* Disallow -it and check if trailing text is not NULL

I noticed that using -it (only read trailing text) crashes since we cannot skip the numerical columns since GMT depends on at least one.  It is OK to use -ot however, to just output the trailing text.  I improved the documentation on this and check for -it to avoid the crash.  I also added checks to ensure a trailing text for a record is not NULL before we access it.

* Update gmt_init.c

Clarify how trailing text is specified.

* Update the rst too

* Add comma in -i synopsis