Avoid the bug in netcdf 4.7.0 by checking path first (#1052)
Instead of calling nc_open with the function gmt_getdatapath, we call gmt_getdatapath separately to examine if it returns NULL.  If it does not then we pass the path we obtain to nc_open.  This is cleaner regardless of whether the netcdf bug is fixed or not.  Closes issue #1045.