Delete incomplete/corrupted files due to Ctrl-C (#1061)
* Delete incomplete/corrupted files due to Ctrl-C

When remove files are copied from the server via Curl, and the user interrupts this via Ctlr-C, we end up with a corrupted file.  Then, if the user tries to access this file (the same day) it will give an error.  This PR responds to #549 by implementing a custom signal handler during the download so that if Ctrl-C happens it is caught and deletes the partial file before exiting.  It excludes Windows for now.

* Add include files!

Forgot...

* Update gmt_remote.c

* Improve names and comments