Enhance trailing text selection by specifying specific word (#1051)
* List file to add extra exe's to the Win installer

* Fix fill issue #1003 (#1004)

* Update PR template to remind contributors to select the correct base branch (#1000)

* Always build branch 6.0 (#1002)

* Fix bug in +/- color filling. Fix #1003

* Bump version to 6.1.0 (#1005)

* Use examples in centimeters instead of inch (#1013)

* Make the gmt_mkdir error messages different (#1033)

* Replace apt-get with apt (#1036)

* VS isdigit in debug mode has an assert C>=0 && c<= 255. Do that here otherwise it crashes in debug mode. (#1049)

* Enhance trailing text selection by specifying specific word

This PR adds an optional word column to the trailing text selection t.  By specifying tword we select one word from the trailing text.  word is the number of the word, starting at 0 for the first word.

* Fix typos

* Add -F+wword for pstext

Since pstext has its own legacy way of reading text, we need to use a new +wword modifier to select a single word from the trailing text as -i is not suitable.

* Switch this option to a local -i

Instead of -F+w, let it be -itword.

* Set minor to 0