Improve pslegend handling of paragraph text
We had two bugs that caused the issue reported in #1096: First, there was a strcpy without checking if the source was not NULL, and second the count for hom many records per paragraphs was not reset to 0 at the start of a new paragraph.  It is possible the first problem was caused by the second bug as well.
In the process I added the hidden capability of developers to compile with DEBUG and use -+ to dump the virtual files passed into psxy and pstext since it is hard to see their contents otherwise.  This PR closes #1096.