Let grdcontour/pscontour be able to have unique pen per contour (#1115)
* Let grdcontour/pscontour be able to have unique pen per contour

THe -Cfile option allows these modules to read specific contours, flag them as C or A, and optionally get a fixed annotation angle.  This PR adds an optional 4th item, a pen, that may also be read from the file.  This allows each contour to be drawn with a unique pen.  If the pen is not present then we default the the one set via -W.

* Minor typos

* No need for the individual_pens variable

Since we have a bool per pen.

* Add some comments