See if using ; under Windows work better for dirs (#1124)
* See if using ; under Windows work better for dirs

In movie, we select the directory separator based on the scripting language, but since this fails on Windows perhaps we should always select ; there.  Give it a try on the animation tests.

* Fix typo

* Use commas for path separators

Since DATA_DIR is just for GMT internal use we should not use OS-specific PATH separators but just use a harmless comma.

* Try the double backslash trick in movie.c

* No more backslashes

No longer use backslash in movie for dir separator.  Replace any incoming backslash in paths with forward slash.

* Amazing...