GMT_Create_Options must always allocate memory (#1134)
When API usage passes a linked list of GMT_OPTION args to GMT_Create_Option it cannot just return that pointer but must duplicate the list and return the head of that duplicate.  That way, gmt_M_free_options can always free the options link and we honor the "const" modifier for that input argument to GMT_Create_Options. See #1132 for context.  Closes #1132.