Fix pstext bug and update example scripts (#1161)
Following #1165, we upgrade 5 scripts to not use the echo command to send a string into pstext.  In the processes I fixed a bug related to the static string given since it needs to be placed in a larger buffer so that substitutinos for color can happen which will make the string longer.