Add -Wfile option in grd2kml for contour overlays (#1117)
* WIP Add -Wfile option in grd2kml for contour overlays

Enable the overlay of select contours, as suggested by #322.  Need some testing first and merging of #1112.

* Fix test

* Update grd2kml.c

Append to PS and pass -R -J args to grdcontour

* Required new API function GMT_Inquire_VirtualFile

Since grcontour can receive a virtual file, we need to find out if CPT or DATASET.  This lead to the addition of this new API function, with wich grd2kml can handle the addition of contours via call to grdcontour.

* Allow -Wpen as well, selecting cpt contours

Works pretty well.