Improve grid region needed for general perspective maps
When we make a general perspective map of a global grid, we must determine the minimum grid region that will be visible.  This is complicated by the fact that the general perspective view may be a clip path composed of straight lines and a curved horizon.  The previous scheme relied on a wesn search along a rectangular path and this failed at times.  The new scheme obtains the clip path and searches along it to determine extreme longitude and latitudes.  This seems to work.  Several origianal PS files needed to be updated as smal lchanges to the resulting regions yield tiny difrerences in grid line coordinates etc.