The auto-region calculation for datasets must be aware of zero range
If all the x-values or all the y-values are equal, then the auto-range detection will give west==east or south==north.  This needs to be prevented.  If equal we scale the min by 0.9 and the max by 1.1, unless they are both zero which we then set to -1 nd + 1.  This closes #1209.