Let the :offset part of pen style be optional, width default 0 (#1219)
If the phase shift in a pen pattern is zero, we should not have to add colon 0 to indicate that - a zero phase should be the default.  We even have a check in gmt_getpen for this and it warns that phase is zero but then it failed to parse the pattern.  Updated the parsing, checking, and docs.