Use proper modifierws to change map width to height
When users specify a map dimension it is by default the map width.  However, we had single characters that could be added at the end to change this: h sets map height, + sets the maximum dimension, anbd - sets the minimum dimensions.  However, these are not proper modifiers.  While still honoring these in a backwards compatibility sense, we now expect +dh, +du, or +dl for the same effects [Default is width +dw].