Let gmt inset provide map width for the next plot
Once we set up a map inset we know the dimensions of the plottable canvas (inset size minus margins).  Thus, if ? is given as map width we substitude the available width instead,  OVer time, I will change this to consider if perhaps the height should be set instead to avoid overflow of the inset.