Let gmt_fill_sytanx and gmt_pen_syntax take a longer option for modifiers (#1319)
Sometimes, we specify a fill color via a modier, e.g., -F+gfill, and then the gmt_fill_syntax funciton fails to show the proper option since it expects a single char.  This PR follows the plan in #1309 and adds that longer option string.  Currently, it is alwas passed as NULL and hence not used, but if you instead pass "F+g" then that is what will show during the message.  COmpiles but not tested.