Save and restore -B settings during inset building
We need to keep the last -B option used before gmt inset starts an inset, so that we can restore that setting as the current last -B setting when the inset ends.  This is needed so that any colormap command can determine if the chosen axis was annotated or not and adjust accordingly.