Old GMT4 -B parsing changed frame selection when seconds were fouund
See issue #1367.  The proble is that when -Ba2sf1sg2s/a2sf1sg2s is parsed we discover that the last s does not mean south and return, but then we have already wiped the side setings to nothing.  Now we check that if we discover the arguments has nothing about the frame then we retain the deafault settings.  Closes #1367.