Clarify tutorial on makecpt and modern mode
The tutorial discusses examples of makecpt which basically are clssic mode commands.  I added text to explain that if you run these commands as part of a modern mode workflow then you either leave off the redirection to a file or your add -H.  I also fixed GMT_tut_9 script in the tutorial (OK in the script).