Warn ir -R with both west and east longitudes are specified
It is very natural to give regions like this to grdcut: -R170E/170W/20/30 but this is internally converted to 170/-170 and fails the w < e test.  We did this for a good reason (if people give -R170/-170 it should be an error) but I think the specific case when (a) it is geogrpahic, (b) W and E are used in setting the longitude, it seems safe to add 360 to east to get a valid domain.  All tests pass so there should not be any side-effects of this.
We ran into this today and could not figure out why -R140E/100W/60S/60N would not work in grdcut...